Make fs block size check independent of architecture.

+++ fs/fs.go
@@ -36,5 +36,5 @@ func blockSizeOf(path string) int64 {
 	if err := syscall.Statfs(path, &stat); err != nil {
 		return 0
 	}
-	return stat.Bsize
+	return int64(stat.Bsize)
 }
